requirements.txt
requests>=2.32.3
blinker==1.8.2
click==8.1.7
Authlib==1.5.2
Flask==3.0.3
flask[async]==3.0.3
flask-socketio>=5.5.1
python-socketio[client]>=5.12.1
Flask-Session>=0.8.0
Flask-Session[redis]>=0.8.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nacos-sdk-python==1.0.0
Werkzeug>=3.0.6
configobj==5.0.8
jproperties==2.1.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
marshmallow==3.21.3
Flask-SQLAlchemy==3.1.1 
sqlalchemy==2.0.32
typing-extensions==4.12.2
protobuf==3.20.1
mysql-connector-python==8.4.0
sqlfluff>=3.1.1
psycopg2-binary>=2.9.10
jsonschema==4.23.0
aws-secretsmanager-caching>=1.1.3
gunicorn==23.0.0
gevent==24.11.1
cryptography==44.0.2
msgpack==1.1.0
#python3-saml>=1.16.0
#xmlsec>=1.3.13
#elasticsearch6>=6.8.2
#elasticsearch7>=7.17.12
#elasticsearch8>=8.16.0
pytest
pytest-cov
pytest-mock